refactor(mvc): use mvc pattern
