+++ README.md
@@ -96,5 +96,5 @@
 ### 구현할 기능 목록
 
 - [x] Domain 패키지로 분리
-- [ ] MVC 패턴 형식으로 분리
+- [x] MVC 패턴 형식으로 분리
