+++ src/main/kotlin/study/racingcar/Application.kt
@@ -1,22 +1,11 @@
 package study.racingcar
 
-import study.racingcar.domain.car.CarFactory
+import study.racingcar.controller.RacingController
 import study.racingcar.domain.generator.RandomNumberGenerator
-import study.racingcar.domain.racing.RacingGame
-import study.racingcar.domain.racing.Winner
-import study.racingcar.io.InputView
-import study.racingcar.io.OutputView
+import study.racingcar.view.InputView
+import study.racingcar.view.OutputView
 
 fun main() {
-    val names = InputView.getCarsName()
-    val round = InputView.getRound()
-
-    val cars = CarFactory.createCars(names)
-    val generator = RandomNumberGenerator()
-
-    val racingGame = RacingGame(cars, round)
-    val racingResult = racingGame.run(generator)
-
-    OutputView.printResult(racingResult)
-    OutputView.printWinners(Winner(cars))
+    val controller = RacingController(InputView(), OutputView(), RandomNumberGenerator())
+    controller.racing()
 }
