+++ src/main/kotlin/study/racingcar/view/InputView.kt
@@ -1,6 +1,6 @@
-package study.racingcar.io
+package study.racingcar.view
 
-object InputView {
+class InputView {
     fun getCarsName(): List<String> {
         println("경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).")
         return readln().split(",")
